Multiple corrections :   - change interface from long to double (error conversions)   - correction error in the order of arguments operands when pop from stack   - add some tests for check these cases

+++ src/main/java/rpn/RpnEvaluator.java
@@ -9,5 +9,5 @@ public interface RpnEvaluator {
      * @param expression the RPN expression
      * @return the result
      */
-    long evaluate(@NonNull @NotNull final String expression);
+    double evaluate(@NonNull @NotNull final String expression);
 }
